// src/api/models/updateDailyFuelStockSummaryRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface UpdateDailyFuelStockSummaryRequest {
  /** ID */
  id: string;
  /** 휘발유 입고 단가(원/L) */
  gasolineIncomePrice: number;
  /** 휘발유 입고량(L) */
  gasolineIncomeVolume: number;
  /** 휘발유 실측재고량(L) */
  gasolineInventory: number;
  /** 경유 입고 단가(원/L) */
  dieselIncomePrice: number;
  /** 경유 입고량(L) */
  dieselIncomeVolume: number;
  /** 경유 실측재고량(L) */
  dieselInventory: number;
  /** 고급유 입고 단가(원/L) */
  premiumIncomePrice: number;
  /** 고급유 입고량(L) */
  premiumIncomeVolume: number;
  /** 고급유 실측재고량(L) */
  premiumInventory: number;
  /**
   * 비고
   * @nullable
   */
  note?: string | null;
}